Addition of other operating systems and *AMP stacks
I have added XAMPP for Windows, MAMP for Apple and LAMP for Linux. 

I have double-checked installation instructions and have not encountered any issues with the instructions given. This however does not include Apple/OSX as I do not currently have a working computer with OSX installed.

Please review and see if these are the type of instructions you are looking for.

I would like to suggest that you add instruction files for each OS which can include the individual steps. These can then also include pictures to assist the user.

Please remove what you do not feel is adequate or deny the changes if you are not satisfied. If you are unsatisfied, please allow me to try again but with more details from your side so I know exactly what you are looking for.